Implemented OpenRead #8
- Added IReadOnly variants of all collections
- Added IDatabase and IReadOnlyDatabase interface, implementation is hidden to the user
- Database acts as "factory"
- Also added ulong as a native data type

+++ IsabelDb/Serializers/UInt32Serializer.cs
@@ -6,7 +6,7 @@ namespace IsabelDb.Serializers
 	internal sealed class UInt32Serializer
 		: ISQLiteSerializer<uint>
 	{
-		public DbType DatabaseType => DbType.UInt32;
+		public DbType DatabaseType => DbType.Int64;
 
 		public object Serialize(uint value)
 		{
